0.0.85: optimize source code
