+++ CHANGELOG
@@ -1,5 +1,9 @@
 Pixel Art Maker
 
+0.0.85 (2019-07-26) <zvezdochiot>
+
+    * optimize source code
+
 0.0.84 (2019-07-26) <zvezdochiot>
 
     * added LICENSE (PDM 1.0)
